mission: mission2 movieList 컴포넌트 분리 movieList 컴포넌트를 분리하여 폴더를 새로 작성했습니다
