+++ mission/chapter02/mission02/src/Page/MovieListPage/index.jsx
@@ -1,27 +1,10 @@
-import useMovieList from "./api/useMovieList";
+import MovieList from "./ui/MovieList";
 import "./style.css";
-const BASE_IMG_URL = "https://image.tmdb.org/t/p/";
-const BASE_IMG_SIZE = "w500";
 
-const MovieList = () => {
-  const [movieList] = useMovieList();
-  if (!movieList) return null;
+const MovieListPage = () => {
   return (
-    <div className="movieList">
-      {movieList.map((elem, index) => {
-        return (
-          <div className="movie" key={index}>
-            <img
-              src={`${BASE_IMG_URL}${BASE_IMG_SIZE}${elem.poster_path}`}
-              alt="movie poster"
-              className="thumbnail"
-            />
-            <div className="thmubnailHoverEffect"></div>
-          </div>
-        );
-      })}
-    </div>
+    <MovieList/>
   );
 };
 
-export default MovieList;
+export default MovieListPage;
