+++ mission/chapter02/mission02/src/Page/MovieListPage/ui/MovieList/style.css
@@ -1,0 +1,28 @@
+.movieList{
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  justify-content: center;
+}
+.movie{
+  position: relative;
+}
+.thumbnail{
+  width: 160px;
+  height: 240px;
+  border-radius: 8px;
+  margin: 0;
+}
+.thmubnailHoverEffect{
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 160px;
+  height: 240px;
+  background-color: gray;
+  opacity: 0;
+  border-radius: 8px;
+}
+.thmubnailHoverEffect:hover{
+  opacity: 0.5;
+}
